+++ Cargo.toml
@@ -26,3 +26,4 @@ windows = { version = "0.52", features = [
 
 [target.'cfg(target_os = "linux")'.dependencies]
 xcb = { version = "1.3", features = ["randr"] }
+smithay-client-toolkit = { version = "0.18.1", default-features = false }
